Move EnergyBalanceModelDefinition together with EnergyBalanceModel and tidy.
