+++ src/HealthGPS/energy_balance_model.h
@@ -1,9 +1,35 @@
 #pragma once
 
-#include "hierarchical_model_types.h"
-#include "random_algorithm.h"
+#include "interfaces.h"
+#include "mapping.h"
 
 namespace hgps {
+
+/// @brief Implements the energy balance model data type
+class EnergyBalanceModelDefinition final {
+  public:
+    EnergyBalanceModelDefinition() = delete;
+
+    /// @brief Initialises a new instance of the EnergyBalanceModelDefinition class
+    /// @param nutrient_coefficients The food group -> nutrient weights
+    /// @throws std::invalid_argument if nutrient coefficients map is empty
+    EnergyBalanceModelDefinition(
+        std::vector<core::Identifier> &&nutrient_list,
+        std::map<core::Identifier, std::map<core::Identifier, double>> &&nutrient_equations);
+
+    /// @brief Gets the list of nutrients used in the model
+    /// @return A vector of nutrient identities
+    const std::vector<core::Identifier> &nutrient_list() const noexcept;
+
+    /// @brief Gets the nutrient coefficients for each food group
+    /// @return A map of nutrient coefficients for each food group
+    const std::map<core::Identifier, std::map<core::Identifier, double>> &
+    nutrient_equations() const noexcept;
+
+  private:
+    std::vector<core::Identifier> nutrient_list_;
+    std::map<core::Identifier, std::map<core::Identifier, double>> nutrient_equations_;
+};
 
 /// @brief Implements the energy balance model type
 ///
@@ -32,4 +58,5 @@ class EnergyBalanceModel final : public HierarchicalLinearModel {
                                                                 Person &entity,
                                                                 int time_year) const;
 };
+
 } // namespace hgps
